Implemented BuzzerMenu

+++ boardcomputer/src/globals.h
@@ -28,11 +28,11 @@ wl_status_t last_status;
 IPAddress last_ip;
 
 struct {
-    int iMotMax = defaultIMotMax;           // [A] Maximum motor current limit
-    int iDcMax = defaultIDcMax;             // [A] Maximum DC Link current limit (This is the final current protection. Above this value, current chopping is applied. To avoid this make sure that I_DC_MAX = I_MOT_MAX + 2A)
-    int nMotMax = defaultNMotMax;           // [rpm] Maximum motor speed limit
-    int fieldWeakMax = defaultFieldWeakMax; // [A] Maximum Field Weakening D axis current (only for FOC). Higher current results in higher maximum speed.
-    int phaseAdvMax = defaultPhaseAdvMax;   // [deg] Maximum Phase Advance angle (only for SIN). Higher angle results in higher maximum speed.
+    int16_t iMotMax = defaultIMotMax;           // [A] Maximum motor current limit
+    int16_t iDcMax = defaultIDcMax;             // [A] Maximum DC Link current limit (This is the final current protection. Above this value, current chopping is applied. To avoid this make sure that I_DC_MAX = I_MOT_MAX + 2A)
+    int16_t nMotMax = defaultNMotMax;           // [rpm] Maximum motor speed limit
+    int16_t fieldWeakMax = defaultFieldWeakMax; // [A] Maximum Field Weakening D axis current (only for FOC). Higher current results in higher maximum speed.
+    int16_t phaseAdvMax = defaultPhaseAdvMax;   // [deg] Maximum Phase Advance angle (only for SIN). Higher angle results in higher maximum speed.
 } settings;
 
 std::array<Controller, 2> controllers{Controller{Serial1}, Controller{Serial2}};
